# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ABG Social Media API",
        "description": "API documentation for the ABG Social Media application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/friends/request/{userId}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Send friend request",
                "operationId": "sendFriendRequest",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id to send request to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request sent"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/friends/accept/{userId}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Accept friend request",
                "operationId": "acceptFriendRequest",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id to accept request from",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request accepted"
                    },
                    "404": {
                        "description": "Friend request not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/friends/reject/{userId}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Reject friend request",
                "operationId": "rejectFriendRequest",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id to reject request from",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request rejected"
                    },
                    "404": {
                        "description": "Friend request not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/friends/{userId}": {
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Remove friend",
                "operationId": "removeFriend",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id to remove from friends",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend removed"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/friends/requests": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get pending friend requests",
                "operationId": "getFriendRequests",
                "responses": {
                    "200": {
                        "description": "List of pending requests"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get friends list",
                "operationId": "getFriends",
                "responses": {
                    "200": {
                        "description": "List of friends"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/friends/cancel/{userId}": {
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Cancel sent friend request",
                "operationId": "cancelFriendRequest",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id to cancel request to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request cancelled"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get user notifications",
                "operationId": "getNotifications",
                "responses": {
                    "200": {
                        "description": "List of notifications"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/notifications/mark-read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "markNotificationsAsRead",
                "responses": {
                    "200": {
                        "description": "Notifications marked as read"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}